Completed all the core functionalities of all the roles. Just Pagination part of fecthing data is remaining. Will complete the core and then define the entire Swagger documentation for the APIs.

+++ src/main/java/com/mukesh/restaurantManagementSystem/dto/request/AssignmentRequestDTO.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Setter
 @Getter
 public class AssignmentRequestDTO {
-    private List<Long> selectedTables;
+    private List<Integer> selectedTables;
     private Long waiterId;
     private boolean temporaryAssignment;
 }
